make control_key optional
this is buggy code, because control_key should be unique and tied to the output device.

we should always require it to be set. this change allows it to be set, but provides a simple/stupid default if it isn't set. this shouldnt break anyone with 1 device, but will definitely break people with more than one.